DataLoader: Fix batch data type for numpy array
The original code is reasonable under purpose without importing numpy. However, in some cases, the user shall generate data set themselves by using Numpy and often the labels are of type `np.float32` which gives True for `type(label).__module__ == numpy`, but it is not the purpose for this line and will raise runtime error, since label is array, it can not be used for `torch.from_numpy()`.